+++ loan/src/main/java/loan/easyloan/entity/UserRequiredInfo.java
@@ -7,7 +7,7 @@ public class UserRequiredInfo {
   private String idCard;
   private String userName;
   private String bankAccount;
-  private long userType;
+  private int userType;
   private String password;
 
 
@@ -47,11 +47,11 @@ public class UserRequiredInfo {
   }
 
 
-  public long getUserType() {
+  public int getUserType() {
     return userType;
   }
 
-  public void setUserType(long userType) {
+  public void setUserType(int userType) {
     this.userType = userType;
   }
